requirements.txt.txt
helper_functions==2.0.11
langchain==0.3.8
langchain_core==0.3.21
langchain_openai==0.2.9
pydantic==2.10.2
python-dotenv==1.0.1
sentence_transformers==3.0.0
streamlit==1.30.0
